fix flooding bug in StatusChecker plugin
TL;DR: This commit fixes each of several bugs which combined effect led vlb to flood when it detected
a severe downtime in its monitored websites.

When using HTTPClient's default timeouts, vlb waits up to ~60 seconds before giving up on a monitored site
(possibly more). However, each 20 seconds, the bot also spawns a new thread which in turns tries to access
the same websites.

But there is worse! In order to prevent flood in case of one of the website's downtime, vlb remembers the last
state of all monitored websites. It compares this memory to each new check result, and warns only if it sees
a difference. Problem was, these memories were composed of either nil or exception objects… which are not
comparable, as evidenced by this irb session:

2.2.3 :001 > a = begin; raise "toto"; rescue => e; e; end
 => #<RuntimeError: toto>
2.2.3 :002 > b = begin; raise "toto"; rescue => e; e; end
 => #<RuntimeError: toto>
2.2.3 :003 > a == b
 => false

So, each time vlb runs the StatusChecker plugin, not only doesn't it wait for a change to appear
before warning again, but it may even not wait for 20s between each warning, depending on how
much time passes before the exception is raised.

The solution was simply:
- to decrease the timeout delay from 60 to 5 seconds;
- to prevent multiple threads from performing their checks concurrently (other threads
    will now give up);
- to store String instance instead of RuntimeError and Addressable::URI instances
    in @last_statuses.